% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatershedScale_Functions.r
\name{cat_shp}
\alias{cat_shp}
\title{Basin Shape Metrics}
\usage{
cat_shp(network, NHDCatchments)
}
\arguments{
\item{network}{a single network extracted from NHDPlusV2}

\item{NHDCatchments}{NHDPlusV2 catchment coverage, Catchment.shp}
}
\value{
sf object of watershed attributed with
width (\code{width}), maximum length (\code{mxln}),
area (\code{area}), parimeter (\code{prmt}),
compactedness coefficient (\code{CmpC}),
circularoty ratio (\code{CrcR}), and elongation
ratio (\code{ElnR}). Where applicable, units = m
}
\description{
Calculates metrics the describe the shape of a networks waterhsed.
}
\details{
Compactness coefficient (CmpC) is defined as the ratio of the
watershed perimeter to the circumference of equivalent circular area.

Elongation ratio (ElnR) is defined as the ratio of diameter of a circle of the same area as
the watershed to the maximum watershed length. The numerical value varies
from 0 (in highly elongated shape) to 1 (in circular shape).

Circulatory ratio (CrcR) is defined as the ratio of watershed area to the area of the circle
having the same perimeter as the watershed perimeter. The numeric value
may vary in between 0 (in line) and 1 (in a circle).
}
