% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComidScale_Functions.r
\name{chn_geom}
\alias{chn_geom}
\title{Channel Shape}
\usage{
chn_geom(network, elevslope)
}
\arguments{
\item{network}{a single network extracted from NHDPlusV2}

\item{elevslope}{NHDPlusV2 elevation slope table, elevslope.dbf"}
}
\value{
data.frame of NHDPlusV2 attributes: slope (\code{SLOPE}), smoothed
maximum \code{MAXELEVSMO} and minimum elevation (\code{MINELEVSMO}) and
COMID length (\code{LENGTHKM}), mean elevation (\code{elev}) and channel sinuosity
(\code{sinuosity}).
}
\description{
Calculates metrics that describe shape of the COMIDs wihtin a network
}
\details{
Elevation is the mean of the smoothed max and min elevation

Sinuosity is COMID length divided by straight line length
}
