% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachScale_Functions.r
\name{DCI_dist}
\alias{DCI_dist}
\title{Distance-Baised Dendritic Connectivity Index}
\usage{
DCI_dist(plen_i, plen_j, dist_ij, mu = 12000)
}
\arguments{
\item{plen_i}{length of patch i}

\item{plen_j}{length of patch j}

\item{dist_ij}{distance separating patch i and patch j}

\item{mu}{mean dispersal distance}
}
\value{
long pairwise patch distances
}
\description{
calculates modified dentritic connectivity
index baised on watercourse distance
}
\details{
index modified from cote et al 2009

median parent offspring dispersal distance is
12km = 12,000m (Comte and Olden 2017)
}
