% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GettingStarted_Functions.r
\name{find_comid}
\alias{find_comid}
\title{Find NHDPlusV2 COMID's

Identifies NHDPlusV2 COMID's within a search radius}
\usage{
find_comid(pts, NHDFlowline, maxdist = 200)
}
\arguments{
\item{pts}{Sampling points}

\item{NHDFlowline}{Flowline coverage from NHDPlusV2}

\item{maxdist}{search radius around points (m)}
}
\value{
list of length points with elements of COMID
within search radius
}
\description{
Find NHDPlusV2 COMID's

Identifies NHDPlusV2 COMID's within a search radius
}
