% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachScale_Functions.r
\name{get_mids}
\alias{get_mids}
\title{INTERNAL FUNCTION identify midpoints}
\usage{
get_mids(coords)
}
\arguments{
\item{coords}{coordinates to identify midpoint}
}
\value{
midpoint returned from set of points
}
\description{
identify mid point from a set of points
used in net sample points
}
\details{
adapted from View(maptools:::getMidpoints)
}
