% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComidScale_Functions.r
\name{get_streamcat}
\alias{get_streamcat}
\title{Network StreamCat Data}
\usage{
get_streamcat(
  comid,
  strcat_path = "Data/StreamCat",
  csvfile = "Lithology",
  vpu = "01"
)
}
\arguments{
\item{comid}{NHDPlusV2 COMID}

\item{strcat_path}{directory containing StreamCat data}

\item{csvfile}{Name of StreamCat File}

\item{vpu}{NHDPlusV2 Vector Processing Unit}
}
\value{
\code{data.frame} of streamCat variables
}
\description{
Extracts StreamCat variables for Network COMID
}
