% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachScale_Functions.r
\name{merge_lines}
\alias{merge_lines}
\title{merge adjacent line segements of the same class}
\usage{
merge_lines(lines, groupName = "grps")
}
\arguments{
\item{lines}{an sf object of LINESTRING}

\item{groupName}{field containing classification}
}
\value{
sf object with adjacent lines of the same class merged together
}
\description{
function takes a series of flowlines and merges adjacent line segments
of the same class
}
