% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComidScale_Functions.r
\name{net_confl}
\alias{net_confl}
\title{Confluence Attributes}
\usage{
net_confl(network, vaa, flow)
}
\arguments{
\item{network}{a single network extracted from NHDPlusV2}

\item{vaa}{NHDPlusV2 value added attributes table, PlusFlowlineVAA.dbf"}

\item{flow}{NHDPlusV2 flow table, PlusFlow.dbf}
}
\value{
data.frame of confluence angle (\code{Confl_angle}),
area ratio (\code{AreaRatio}), total upstream area \code{Confl_AREA},
order of the tributary (\code{Confl_order}), and tributary classification
\code{Confl_class}
}
\description{
Calculates metrics related to confluences in a network
}
\details{
Lambert Conformal Cone (crs = 102004) used to
preserve angles Seybold et al. (2017) is no longer available

in some cases tributary angle is NAN because average tributary direction is near verticle.
the function cannot determine starts/end of line needed to calculate direction position
vector direction. this could be could fix this by changing the maximum verticies

Complex confluences >2 tributaries are incorrect.

Confluence angle is the angle two tributarys meet. Calculated as the intersection between
two slopes orthoganal regression slopes reflecting each tributary's average direction

Area ratio is the ratio of the tributary catchment area to the mainstem catchment area. Mainstem
is identified as tributary with the largest catchment area

confluence area is the total area upstream of the confluence

confluence order is the stream order directly downstream of the confluence

confluence class is the stream order of each tributary in the confluence
}
