% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatershedScale_Functions.r
\name{net_delin}
\alias{net_delin}
\title{Network Delineation}
\usage{
net_delin(comid, flow = flow, vaa = vaa)
}
\arguments{
\item{comid}{COMID at network outlet}

\item{flow}{NHDPlusV2 flow table, PlusFlow.dbf}

\item{vaa}{NHDPlusV2 value added attributes table, PlusFlowlineVAA.dbf"}
}
\value{
vector of upstream COMIDs
}
\description{
Identifies flowlines upstream from a COMID
}
