% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachScale_Functions.r
\name{net_sample}
\alias{net_sample}
\title{Create Network Reaches, Points and Transects}
\usage{
net_sample(
  network,
  vaa,
  n = 5,
  what = c("reaches", "midpoints", "transects"),
  NHDCatchments = NULL,
  keep.divergent = F
)
}
\arguments{
\item{network}{a single network extracted from NHDPlusV2}

\item{vaa}{NHDPlusV2 value added attributes table, PlusFlowlineVAA.dbf"}

\item{n}{number of reaches per network COMID}

\item{what}{one of c("reaches", "midpoints", "transects")}

\item{NHDCatchments}{NHDPlusV2 catchment coverage, Catchment.shp, required if what = "transects"}

\item{keep.divergent}{logical keep divergent flow paths}
}
\value{
sf object of either lines or points. Transects are perpendicular to the reach
}
\description{
Splits network COMIDs into reaches, identifies midpoints
and creates lateral transects
}
