% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatershedScale_Functions.r
\name{net_shp}
\alias{net_shp}
\title{Network Shape Metrics}
\usage{
net_shp(network, vaa)
}
\arguments{
\item{network}{a single network extracted from NHDPlusV2}

\item{vaa}{NHDPlusV2 value added attributes table, PlusFlowlineVAA.dbf"}
}
\value{
data.frame with network stream order (\code{Ord}),
Mainstem Length (\code{Ord}), total length (\code{TL}), Drainage
Density (\code{Dd}), Horton Laws: bifurcation ratio (\code{Rb}),
length Ratio (\code{Rl}) and area Ratio (\code{Ord})
}
\description{
Calculates metrics the describe the shape of a network
}
\details{
Horton ratios estimated as semilog relationships
}
