% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachScale_Functions.r
\name{patch_dist}
\alias{patch_dist}
\title{average distance between patches}
\usage{
patch_dist(graph, net_HGP, groupName, patch = "2")
}
\arguments{
\item{graph}{result from sf_todygraph}

\item{net_HGP}{a classified stream network}

\item{groupName}{name of field containing classification result}

\item{patch}{calss value}
}
\value{
long pairwise patch distances
}
\description{
embedded within class metrics
}
