% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComidScale_Functions.r
\name{pos_vec}
\alias{pos_vec}
\title{Average line direction}
\usage{
pos_vec(p, maxvert = nrow(p), diagnostics = T)
}
\arguments{
\item{p}{an points of an sf LINESTRING}

\item{maxvert}{maximum number of line verticies used to fit regression}

\item{diagnostics}{logical used for plotting}
}
\value{
position vector of a line unless diagnostics = T
}
\description{
Calculates average direction of a line used for concluence angle
}
\details{
this function uses orthogonal regression to find the
average direction of a series of points and returns
the position vector.

max vert is the maximum number of verticies to fit average line

in some instances the function will return NAN when
line is near verticle. Could fix this by changing the
maximum verticies
}
