% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReachScale_Functions.r
\name{sf_to_tidygraph}
\alias{sf_to_tidygraph}
\title{sf to tidygraph}
\usage{
sf_to_tidygraph(x, directed = TRUE)
}
\arguments{
\item{x}{sf network}

\item{directed}{logical}
}
\value{
super useful graph
}
\description{
creates tidy graph from sf object
}
\details{
adapted from https://www.r-spatial.org/r/2019/09/26/spatial-networks.html embedded within class metrics
}
