% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GettingStarted_Functions.r
\name{StreamCat_download}
\alias{StreamCat_download}
\title{StreamCat Region Download}
\usage{
StreamCat_download(
  path = getwd(),
  vpu = "17",
  filename = "NRSA_PredictedBioCondition"
)
}
\arguments{
\item{path}{Directory for downloaded files}

\item{vpu}{NHDPlusV2 Vector Processing Unit}

\item{filename}{name of streamcat file to download}
}
\value{
StreamCat data files are downloaded to "StreamCat" directory
}
\description{
Downloads zip files for VPU from StreamCat webpage
\url{https://www.epa.gov/national-aquatic-resource-surveys/streamcat}
}
