% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComidScale_Functions.r
\name{vec_angle}
\alias{vec_angle}
\title{Intersection Angle}
\usage{
vec_angle(pos1, pos2, deg = T)
}
\arguments{
\item{pos1}{position vector of line 1}

\item{pos2}{position vector of line 2}

\item{deg}{logical return angle in degrees}
}
\value{
intersection angle of two positon vectors
}
\description{
Intersection Angle
}
